+++ packages/icejs/src/getBuiltInPlugins.ts
@@ -31,12 +31,8 @@ const getBuiltInPlugins: IGetBuiltInPlugins = (userConfig) => {
     // react base plugin
     'build-plugin-react-app',
 
-    // for ice/miniapp plugins
-    'build-plugin-miniapp',
-
     // for ice/react plugins
     'build-plugin-ice-router',
-    'build-plugin-ice-helpers',
     'build-plugin-ice-logger',
     'build-plugin-ice-config',
     'build-plugin-ice-mpa',