+++ packages/plugin-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.0
+
+- [feat] migrate runtime api
+
 ## 1.8.1
 
 - [fix] `userConfig` ts type in `config.ts.ejs`