+++ packages/plugin-ice-ssr/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 3.0.0
+
+- [feat] resolve runtime in server module
+- [feat] update loader options for webpack 5
+
 ## 2.2.0
 
 - [feat] support setting getInitialProps with router configuration